# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Cards API
  description: REST API для управления банковскими картами
  contact:
    name: Bank Cards Team
    email: support@bankcards.com
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /user/transfers:
    post:
      tags:
        - card-controller
      summary: Перевод между своими картами
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
        required: true
      responses:
        '200':
          description: Перевод успешно выполнен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Недостаточно средств, карта неактивна или перевод на ту же карту
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Одна из карт принадлежит другому пользователю
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/cards:
    get:
      tags:
        - card-controller
      summary: Получить свои карты
      description: >-
        Возвращает список карт текущего пользователя с пагинацией и поиском по
        номеру
      operationId: getMyCards
      parameters:
        - name: search
          in: query
          description: 'Поиск по маске номера карты (например: 1234)'
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          description: 'Пагинация и сортировка. Пример: page=0&size=10&sort=createdAt,desc'
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Список карт успешно получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          description: Неверные параметры запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - card-controller
      summary: Создать новую карту
      description: Создает новую карту для текущего пользователя
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
        required: true
      responses:
        '200':
          description: Карта успешно создана
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Неверные данные карты (имя или срок действия)
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
  /user/cards/{cardId}/block:
    post:
      tags:
        - card-controller
      summary: Заблокировать карту
      description: Переводит карту в статус BLOCKED. Доступна только для своих карт
      operationId: blockCard
      parameters:
        - name: cardId
          in: path
          description: ID карты для блокировки
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Карта успешно заблокирована
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Карта уже заблокирована
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Карта принадлежит другому пользователю
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
  /user/cards/{cardId}/activate:
    post:
      tags:
        - card-controller
      summary: Активировать карту
      description: Переводит карту в статус ACTIVE. Доступна только для своих карт
      operationId: activateCard
      parameters:
        - name: cardId
          in: path
          description: ID карты для активации
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Карта успешно активирована
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Карта уже активна или истек срок
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '403':
          description: Карта принадлежит другому пользователю
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Карта не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
...
  /admin/users/{userId}/cards:
    get:
      tags:
        - Admin Controller
      summary: Получить карты пользователя
      description: >-
        Возвращает список карт указанного пользователя с пагинацией и поиском.
        Только для ADMIN
      operationId: getUserCards
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
        - name: search
          in: query
          description: 'Поиск по маске номера карты (например: 1234)'
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          description: 'Пагинация: page, size, sort (например: sort=createdAt,desc)'
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Список карт успешно получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '400':
          description: Неверный ID пользователя или параметры запроса
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен - требуется роль ADMIN
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Admin Controller
      summary: Создать карту для пользователя
      description: >-
        Создает новую карту для указанного пользователя. Номер генерируется
        автоматически. Только для ADMIN
      operationId: createCardForUser
      parameters:
        - name: userId
          in: path
          description: ID пользователя
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardRequest'
        required: true
      responses:
        '200':
          description: Карта успешно создана
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardResponse'
        '400':
          description: Неверные данные карты
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /user/cards/{cardId}/balance:
    get:
      tags:
        - card-controller
      summary: Получить баланс карты
      operationId: getCardBalance
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Баланс успешно получен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Карта принадлежит другому пользователю
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Карта не найдена
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: Время возникновения ошибки
          format: date-time
        status:
          type: integer
          description: HTTP статус код
          format: int32
          example: 400
        error:
          type: string
          description: Тип ошибки
          example: Bad Request
        message:
          type: string
          description: Сообщение об ошибке
          example: 'User not found with id: 1'
        path:
          type: string
          description: Путь запроса
          example: /api/users/1
        validationErrors:
          type: object
          additionalProperties:
            type: string
            description: Детальные ошибки валидации (поле -> сообщение)
            example: '{"username":"Username is required","email":"Invalid email format"}'
          description: Детальные ошибки валидации (поле -> сообщение)
          example:
            username: Username is required
            email: Invalid email format
      description: Ответ с информацией об ошибке
...
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          description: ID карты
          format: int64
          example: 1
        cardNumberMasked:
          type: string
          description: Маскированный номер карты
          example: '**** **** **** 1234'
        cardHolderName:
          type: string
          description: Имя владельца на карте
          example: IVAN PETROV
        expirationDate:
          type: string
          description: Срок действия
          format: date
          example: '2025-12-31'
        status:
          type: string
          description: Статус карты
          example: ACTIVE
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
        balance:
          type: number
          description: Баланс
          example: 15000.5
        userId:
          type: integer
          description: ID владельца карты
          format: int64
          example: 1
      description: Ответ с данными банковской карты
    TransferRequest:
      required:
        - amount
        - fromCardId
        - toCardId
      type: object
      properties:
        fromCardId:
          type: integer
          description: ID карты, с которой списываются средства
          format: int64
          example: 1
        toCardId:
          type: integer
          description: ID карты, на которую зачисляются средства
          format: int64
          example: 2
        amount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
          description: Сумма перевода (минимум 0.01)
          example: 1000.5
      description: Запрос на перевод средств между картами
    TransferResponse:
      type: object
      properties:
        message:
          type: string
          description: Сообщение о результате
          example: Transfer completed successfully
        transferId:
          type: integer
          description: ID перевода
          format: int64
          example: 1
        amount:
          type: number
          description: Сумма перевода
          example: 1000.5
        fromCardId:
          type: integer
          description: ID карты отправителя
          format: int64
          example: 1
        toCardId:
          type: integer
          description: ID карты получателя
          format: int64
          example: 2
        timestamp:
          type: string
          description: Время перевода
          format: date-time
      description: Ответ на перевод средств
    CardRequest:
      required:
        - cardHolderName
        - expirationDate
      type: object
      properties:
        cardHolderName:
          maxLength: 100
          minLength: 2
          type: string
        expirationDate:
          type: string
          format: date
      description: Данные новой карты
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          description: Список карт на текущей странице
          items:
            $ref: '#/components/schemas/CardResponse'
        number:
          type: integer
          description: Номер текущей страницы (начиная с 0)
          format: int32
          example: 0
        size:
          type: integer
          description: Размер страницы
          format: int32
          example: 10
        totalElements:
          type: integer
          description: Общее количество элементов
          format: int64
          example: 25
        totalPages:
          type: integer
          description: Общее количество страниц
          format: int32
          example: 3
        hasNext:
          type: boolean
          description: Есть ли следующая страница
          example: true
        hasPrevious:
          type: boolean
          description: Есть ли предыдущая страница
          example: false
        first:
          type: boolean
          description: Первая ли это страница
          example: true
        last:
          type: boolean
          description: Последняя ли это страница
          example: false
      description: Страница с картами
    BalanceResponse:
      type: object
      properties:
        cardId:
          type: integer
          description: ID карты
          format: int64
          example: 1
        balance:
          type: number
          description: Текущий баланс
          example: 1500.5
      description: Ответ с балансом карты
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT